ffs/requirements.txt
pytest
Django==5.0.1
django-ninja==1.1.0 ##
psycopg2-binary==2.9.11
pyjwt==2.10.1